Move Spell to common
