+++ client/Library/SpellLibrary.ts
@@ -3,7 +3,7 @@ import * as ko from "knockout";
 import { ServerListing } from "../../common/Listable";
 import { concatenatedStringRegex } from "../../common/Toolbox";
 import { AccountClient } from "../Account/AccountClient";
-import { Spell } from "../Spell/Spell";
+import { Spell } from "../../common/Spell";
 import { Store } from "../Utility/Store";
 import { Listing, ListingOrigin } from "./Listing";
