+++ server/server.ts
@@ -1,6 +1,6 @@
 import express = require("express");
 import socketIO = require("socket.io");
-import { Spell } from "../client/Spell/Spell";
+import { Spell } from "../common/Spell";
 import { StatBlock } from "../common/StatBlock";
 import * as DB from "./dbconnection";
 import LaunchServer from "./launchserver";
